Rename classes with "GillyCarpetAddons" to "CarpetAddonsNotFound"
